requirements.txt
czifile>=2019.7.2
h5py>=2.10.0
matplotlib>=3.1.3
nibabel>=3.1.0
numpy>=1.18.1
pyyaml>=5.3.1
scikit-image>=0.16.2
scipy>=1.4.1
simpleitk>=2.0.0rc1.post285
pandas>=1.0.4
